moved generateFuzzyCombinedChange.ts
